# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API E-Commerce UAS",
        "description": "Dokumentasi lengkap API untuk integrasi Frontend Mobile (Flutter).",
        "contact": {
            "email": "hifadlibilal@sistekin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/cart": {
            "get": {
                "tags": [
                    "Cart (Buyer)"
                ],
                "summary": "Lihat Keranjang Belanja",
                "operationId": "294e093bf29c111c01d8402121fb4f43",
                "responses": {
                    "200": {
                        "description": "Detail Keranjang",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "description": "ID Cart Item",
                                                                "type": "integer"
                                                            },
                                                            "quantity": {
                                                                "type": "integer"
                                                            },
                                                            "product": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer"
                                                                    },
                                                                    "name": {
                                                                        "type": "string"
                                                                    },
                                                                    "price": {
                                                                        "type": "integer"
                                                                    },
                                                                    "stock": {
                                                                        "description": "Sisa stok saat ini",
                                                                        "type": "integer"
                                                                    },
                                                                    "image": {
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cart (Buyer)"
                ],
                "summary": "Tambah Barang ke Keranjang",
                "operationId": "17edc07162b12ee962c37746abcf34a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil masuk keranjang"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/item/{itemId}": {
            "put": {
                "tags": [
                    "Cart (Buyer)"
                ],
                "summary": "Update Qty (+/-)",
                "operationId": "7f3fa964fa61418b0cdf813c3647415a",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Qty berhasil diupdate"
                    },
                    "400": {
                        "description": "Stok tidak mencukupi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart (Buyer)"
                ],
                "summary": "Hapus Item dari Keranjang",
                "operationId": "40390fca795c3602b781a99fe66ec926",
                "parameters": [
                    {
                        "name": "itemId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/checkout": {
            "post": {
                "tags": [
                    "Order & Payment"
                ],
                "summary": "Checkout (Buat Invoice Xendit)",
                "operationId": "6ed91804beb738578b8d4c98f822dbbf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shipping_service": {
                                        "type": "string",
                                        "example": "REG"
                                    },
                                    "shipping_cost": {
                                        "type": "integer",
                                        "example": 20000
                                    },
                                    "courier": {
                                        "type": "string",
                                        "example": "jne"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sukses, dapat Link Payment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payment_url": {
                                            "type": "string",
                                            "example": "https://checkout-staging.xendit.co/..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Order & Payment"
                ],
                "summary": "History Order Saya",
                "operationId": "325fd3d55bbe9bb39645269928fa4983",
                "responses": {
                    "200": {
                        "description": "List History Order"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan Token Bearer di sini untuk mengakses endpoint yang dilindungi.",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}